+++ templates/full/javascript/src/react/main.jsx
@@ -1,36 +1,36 @@
-var react = require('react');
+import React from 'react';
+import ReactDOM from 'react-dom';
 
 // View that displays a button to call the server
-var NameLoaderView = react.createClass({
-  render: function() {
-    return (
-      <div>
-        <h1>{this.props.staticName}</h1>
-        <p>Hello {this.props.staticName}</p>
-        <input type="button" value="Get app name" onClick={this.props.handleClick} />
-        <p>{this.props.appName}</p>
-      </div>
-    );
-  }
-});
+var NameLoaderView = (props) => (
+  <div>
+    <h1>{props.staticName}</h1>
+    <p>Hello {props.staticName}</p>
+    <input type="button" value="Get app name" onClick={props.handleClick} />
+    <p>{props.appName}</p>
+  </div>
+);
 
 // React component that handles the button click
-var NameLoader = react.createClass({
-  getInitialState: function () {
-    return { appName: '' };
-  },
-  handleClick: function () {
+class NameLoader extends React.Component {
+  constructor(props) {
+    super(props);
+    
+    this.state = { appName: '' };
+    this.handleClick = this.handleClick.bind(this);
+  }
+  handleClick() {
     var xhr = new XMLHttpRequest();
     xhr.open('GET', '/api', true);
-    xhr.onload = function () {
+    xhr.onload = () => {
       this.setState({ appName: xhr.responseText });
-    }.bind(this);
+    }
     xhr.send();
-  },
-  render: function () {
+  }
+  render() {
     return <NameLoaderView appName={this.state.appName} staticName={this.props.staticName} handleClick={this.handleClick} />;
   }
-});
+}
 
 // Tell react to render the component
-react.render(<NameLoader staticName="<%= appname %>" />, document.getElementById('content'));
+ReactDOM.render(<NameLoader staticName="<%= appname %>" />, document.getElementById('content'));