+++ templates/full/typescript/src/react/main.tsx
@@ -1,4 +1,5 @@
 import React = require('react');
+var ReactDOM : __ReactDom = require('react-dom');
 
 class NameLoaderState {
   public appName: string;
@@ -16,14 +17,12 @@ class NameLoaderViewProps extends NameLoaderProps {
 // View that displays a button to call the server
 class NameLoaderView extends React.Component<NameLoaderViewProps, {}> {
   render() {
-    return (
-      <div>
-        <h1>{this.props.staticName}</h1>
-        <p>Hello {this.props.staticName}</p>
-        <input type="button" value="Get app name" onClick={this.props.handleClick} />
-        <p>{this.props.appName}</p>
-        </div>
-    );
+    return <div>
+      <h1>{this.props.staticName}</h1>
+      <p>Hello {this.props.staticName}</p>
+      <input type="button" value="Get app name" onClick={this.props.handleClick} />
+      <p>{this.props.appName}</p>
+      </div>;
   }
 }
 
@@ -49,4 +48,4 @@ class NameLoader extends React.Component<NameLoaderProps, NameLoaderState> {
 }
 
 // Tell react to render the component
-React.render(<NameLoader staticName="<%= appname %>" />, document.getElementById('content'));
+ReactDOM.render(<NameLoader staticName="<%= appname %>" />, document.getElementById('content'));